% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bulk.R
\name{add_noise}
\alias{add_noise}
\title{Add noise to simulated count data}
\usage{
add_noise(sim_pseudo, sd1 = 100, sd2 = 0.01)
}
\arguments{
\item{sim_pseudo}{An integer count matrix with genes in rows and cell
subclasses typically generated by \code{\link[=simulate_bulk]{simulate_bulk()}}.}

\item{sd1}{Standard deviation of noise added to counts.}

\item{sd2}{Standard deviation of noise added to log2(counts+1).}
}
\value{
An integer count matrix with genes in rows and cell subclasses in
columns.
}
\description{
Gaussian noise can be added to the simulated count matrix in 2 ways which can
be combined. \code{sd1} controls addition of simple Gaussian noise to counts using
\code{rnorm} with sd specified by \code{sd1}. If \code{sd2} is >0, counts are converted
using log2+1 and Gaussian noise added, followed by conversion back to count
scale. Negative values are converted to 0. \code{sd1} affects low expressed genes
and hardly affects high expressed genes. \code{sd2} affects all genes irrespective
of expression level.
}
