% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_library_size.R
\name{adjust_library_size}
\alias{adjust_library_size}
\title{Adjust count matrix by library size}
\usage{
adjust_library_size(x)
}
\arguments{
\item{x}{Read count matrix with genes in rows and samples in columns.}
}
\value{
Matrix of adjusted read counts
}
\description{
Simple tool for adjusting raw count matrix by total library size. Library
size is calculated as column sums and columns are scaled to the median total
library size.
}
