% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMarkers.R
\name{cellMarkers}
\alias{cellMarkers}
\title{Identify cell markers}
\usage{
cellMarkers(
  scdata,
  bulkdata = NULL,
  subclass,
  cellgroup = NULL,
  nsubclass = 5,
  ngroup = 5,
  expfilter = 1,
  noisefilter = 2,
  noisefraction = 0.25,
  min_cells = 10,
  big = NULL,
  verbose = TRUE,
  sliceSize = 5000L,
  cores = 1L
)
}
\arguments{
\item{scdata}{Single-cell data matrix with genes in rows and cells in
columns.}

\item{bulkdata}{Optional data matrix containing bulk RNA-Seq data with genes
in rows. This matrix is only used for its rownames, to ensure that cell
markers are selected from genes in the bulk dataset.}

\item{subclass}{Vector of cell subclasses matching the columns in \code{scdata}}

\item{cellgroup}{Optional grouping vector of major cell types matching the
columns in \code{scdata}. \code{subclass} is assumed to contain subclasses which are
subsets within \code{cellgroup} overarching classes.}

\item{nsubclass}{Number of genes to select for each single cell subclass.}

\item{ngroup}{Number of genes to select for each cell group.}

\item{expfilter}{Genes whose maximum mean expression on log2 scale per cell
type are below this value are removed and not considered for the signature.}

\item{noisefilter}{Sets an upper bound for \code{noisefraction} cut-off below
which gene expression is set to 0. Essentially gene expression above this
level must be retained in the signature. Setting this higher can allow more
suppression via noisefraction and can favour more highly expressed genes.}

\item{noisefraction}{Numeric value. Maximum mean log2 gene expression across
cell types is calculated and values in celltypes below this fraction are
set to 0. Set in conjunction with \code{noisefilter.} Note: if this is set too
high (too close to 1), it can have a deleterious effect on deconvolution.}

\item{min_cells}{Numeric value specifying minimum number of cells in a
subclass category. Subclass categories with fewer cells will be ignored.}

\item{big}{Logical whether to invoke matrix slicing to handle big matrices.}

\item{verbose}{Logical whether to show messages.}

\item{sliceSize}{Integer, number of rows of \code{x} to use in each slice if
\code{big = TRUE}.}

\item{cores}{Integer, number of cores to use for parallelisation using
\code{mclapply()}. Parallelisation is not available on windows. Warning:
parallelisation has increased memory requirements. See \code{\link[=scmean]{scmean()}}.}
}
\value{
Returns list object containing \code{best_angle}, a list of genes ranked by lowest
angle and highest maximum expression in a cell type; \code{genemeans}, matrix of
mean log2+1 gene expression with genes in rows and cell types in columns;
\code{genemeans_filters}, matrix of gene expression following noise reduction.
}
\description{
Uses geometric method based on vector dot product to identify genes which are
the best markers for individual cell types.
}
