% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMarkers.R
\name{collapse_group}
\alias{collapse_group}
\title{Collapse groups in cellMarkers object}
\usage{
collapse_group(mk, groups, weights = NULL)
}
\arguments{
\item{mk}{A 'cellMarkers' class object.}

\item{groups}{Character vector of groups to be collapsed. The collapsed group
retains the name of the 1st element.}

\item{weights}{Optional vector of weights for calculating the mean gene
expression across groups. If left as \code{NULL} weights are determined by the
total cell count in each group.}
}
\value{
An updated cellMarkers class object.
}
\description{
Experimental function for collapsing groups in a cellMarkers objects.
}
