% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_heatmap.R
\name{comp_heatmap}
\alias{comp_heatmap}
\title{Compensation heatmap}
\usage{
comp_heatmap(x, cell_table = NULL, ...)
}
\arguments{
\item{x}{object of class 'deconv' or a matrix of compensation values.}

\item{cell_table}{optional grouping vector to separate the heatmap rows and
columns into groups.}

\item{...}{optional arguments passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}}
}
\value{
No return value. Draws a ComplexHeatmap.
}
\description{
Plots a heatmap of the compensation matrix for cell subclasses using
ComplexHeatmap.
}
