% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_similarity.R
\name{cos_similarity}
\alias{cos_similarity}
\title{Gene signature cosine similarity matrix}
\usage{
cos_similarity(x, use_filter = NULL)
}
\arguments{
\item{x}{Either a matrix or a 'cellMarkers' class or 'deconv' class object.}

\item{use_filter}{Logical whether to use filtered gene signature.}
}
\value{
A symmetric similarity matrix.
}
\description{
Computes the cosine similarity matrix from the gene signature matrix of a
\code{cellMarkers} object or any matrix. Note that this function computes cosine
similarity between matrix columns, unlike \code{\link[=dist]{dist()}} which computes the
distance metric between matrix rows.
}
