% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute}
\alias{deconvolute}
\title{Deconvolute bulk RNA-Seq using single-cell RNA-Seq signature}
\usage{
deconvolute(
  mk,
  test,
  log = TRUE,
  comp_amount = 1,
  group_comp_amount = 0,
  equal_weight = FALSE,
  adjust_comp = TRUE,
  use_filter = TRUE,
  convert_bulk = TRUE
)
}
\arguments{
\item{mk}{object of class 'cellMarkers'}

\item{test}{matrix of bulk RNA-Seq to be deconvoluted. We recommend raw
counts as input, but normalised data can be provided, in which case set
\code{log = FALSE}.}

\item{log}{Logical, whether to apply log2 +1 to count data in \code{test}. Set to
\code{FALSE} if prenormalised bulk RNA-Seq data is provided.}

\item{comp_amount}{either a single value from 0-1 for the amount of
compensation or a numeric vector with the same length as the number of cell
subclasses to deconvolute.}

\item{group_comp_amount}{either a single value from 0-1 for the amount of
compensation for cell group analysis or a numeric vector with the same
length as the number of cell groups to deconvolute.}

\item{equal_weight}{logical, whether the gene signature matrix is scaled to
equalise the effects of each gene.}

\item{adjust_comp}{logical, whether to optimise \code{comp_amount} to prevent
negative cell proportion projections.}

\item{use_filter}{logical, whether to use denoised signature matrix}

\item{convert_bulk}{logical, whether to convert bulk RNA-Seq to scRNA-Seq
scaling.}
}
\value{
a list object of class 'deconv' containing the cell proportions of
each cell subclass in each sample in element \code{subclass}, and the cell
proportions of each cell group in element \code{group}.
}
\description{
Deconvolution of bulk RNA-Seq using vector projection method with optional
compensation for spillover.
}
