% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Diagnostics for cellMarker signatures}
\usage{
diagnose(mk)
}
\arguments{
\item{mk}{A 'cellMarkers' class object.}
}
\value{
No return value. Displays information about the cellMarkers
signature showing cells subclasses with weak signatures and diagnostic
information including which cell subclasses each problematic signature
spills into.
}
\description{
Identifies cell subclasses or groups with weak signatures.
}
