% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_bulk.R
\name{fix_bulk}
\alias{fix_bulk}
\title{Fix in missing genes in bulk RNA-Seq matrix}
\usage{
fix_bulk(mk, bulk)
}
\arguments{
\item{mk}{object of class 'cellMarkers'. See \code{\link[=cellMarkers]{cellMarkers()}}.}

\item{bulk}{matrix of bulk RNA-Seq}
}
\value{
Expanded bulk matrix with extra rows for missing genes, filled with
zeros.
}
\description{
Fills in missing genes in a bulk RNA-Seq matrix based on the gene signature
of a 'cellMarkers' objects. Signature is taken from both the subclass gene
set and group gene set.
}
