% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMarkers.R
\name{fix_group}
\alias{fix_group}
\title{Fix cellMarkers signature with no cell groups}
\usage{
fix_group(mk, lab)
}
\arguments{
\item{mk}{A 'cellMarkers' class object.}

\item{lab}{Character value to label the overarching group.}
}
\value{
A 'cellMarkers' class list object in which the elements \code{cell_table}
and \code{groupmeans} have been updated.
}
\description{
This function is designed to fix \code{cellMarkers} objects which were not created
with a \code{cellgroup} vector and therefore have no cell grouping categories.
This can cause issues during merging of cellMarkers objects.
}
\seealso{
\code{\link[=mergeMarkers]{mergeMarkers()}}
}
