% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_angle.R
\name{gene_angle}
\alias{gene_angle}
\title{Vector based best marker selection}
\usage{
gene_angle(genemeans)
}
\arguments{
\item{genemeans}{matrix of mean gene expression with genes in rows and
celltypes, tissues or subclasses in columns.}
}
\value{
a list whose length is the number of columns in genemeans, with each
element containing a dataframe with genes in rows, sorted by best marker
status as determined by minimum vector angle and highest maximum gene
expression per celltype/tissue.
}
\description{
Core function which takes a matrix of mean gene expression (assumed to be
log2 transformed to be more Gaussian). Mean gene expression per gene is
scaled to a unit hypersphere assuming each gene represents a vector in space
with dimensions representing each cell subclass/group.
}
