% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bulk.R
\name{generate_cells}
\alias{generate_cells}
\title{Generate random cell number samples}
\usage{
generate_cells(object, n)
}
\arguments{
\item{object}{A 'cellMarkers' class object}

\item{n}{Integer value for the number of samples to generate}
}
\value{
An integer matrix with \code{n} rows, with columns for each cell
subclasses in \code{object}, representing cell counts for each cell subclass.
Designed to be passed to \code{\link[=simulate_bulk]{simulate_bulk()}}.
}
\description{
Used for simulating pseudo-bulk RNA-Seq from a 'cellMarkers' object.
}
\seealso{
\code{\link[=simulate_bulk]{simulate_bulk()}}
}
