% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudent.R
\name{kappa.deconv}
\alias{kappa.deconv}
\title{Compute condition number of deconvolution model}
\usage{
\method{kappa}{deconv}(x, ...)
}
\arguments{
\item{x}{a 'deconv' class object}

\item{...}{arguments passed to \code{\link[=kappa]{kappa()}}}
}
\value{
The condition number.
}
\description{
Computes the condition number of the spillover matrix from a deconvolution
model.
}
