% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_similarity.R
\name{max_similarity}
\alias{max_similarity}
\title{Maximum similarity between cell types}
\usage{
max_similarity(x, ...)
}
\arguments{
\item{x}{Either a matrix or a 'cellMarkers' class or 'deconv' class object.}

\item{...}{Optional arguments passed to \code{\link[=cos_similarity]{cos_similarity()}}.}
}
\value{
A vector of similarity results.
}
\description{
Computes maximum similarity between cell types based on the gene signature
matrix.
}
\seealso{
\code{\link[=diagnose]{diagnose()}} which outputs a more comprehensive pairwise comparison
of similarity between cell subclasses.
}
