% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_map.R
\name{plot.qqmap}
\alias{plot.qqmap}
\title{Quantile-quantile plot}
\usage{
\method{plot}{qqmap}(x, points = TRUE, ...)
}
\arguments{
\item{x}{A 'qqmap' class object created by \code{\link[=quantile_map]{quantile_map()}}.}

\item{points}{Logical whether to show quantile points.}

\item{...}{Optional plotting parameters passed to \code{\link[=plot]{plot()}}.}
}
\value{
No return value. Produces a QQ plot using base graphics with a red
line showing the conversion function.
}
\description{
Produces a QQ plot showing the conversion function from the first dataset to
the second.
}
