% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_deconv.R
\name{plot_biasvar}
\alias{plot_biasvar}
\title{Plot bias-variance decomposition}
\usage{
plot_biasvar(result, subclass = 1, xvar = colnames(result)[1])
}
\arguments{
\item{result}{Dataframe of tuning results generated by \code{\link[=tune_deconv]{tune_deconv()}}.}

\item{subclass}{Character value specifying which cell subclass to plot, or
numeric selecting the subclass by factor level.}

\item{xvar}{Character value specifying column in \code{result} to vary along the x
axis.}
}
\value{
ggplot2 line plot.
}
\description{
Produces a ggplot2 plot of total error (MSE) against bias and variance based
on results generated by \code{\link[=tune_deconv]{tune_deconv()}}. Primarily useful for examining the
effect of altering compensation or lambda.
}
