% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp.R
\name{plot_comp}
\alias{plot_comp}
\title{Plot compensation analysis}
\usage{
plot_comp(x, ...)
}
\arguments{
\item{x}{An object of class 'deconv' generated by \code{\link[=deconvolute]{deconvolute()}}.}

\item{...}{Optional graphical arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{
No return value, plots the effect of varying compensation on minimum
subclass output for each cell subclass.
}
\description{
Plots the effect of varying compensation from 0 to 1 for each cell subclass,
examining the minimum subclass output result following a call to
\code{\link[=deconvolute]{deconvolute()}}. For this function to work, the argument \code{plot_comp} must be
set to \code{TRUE} during the call to \code{\link[=deconvolute]{deconvolute()}}.
}
