% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_deconv.R
\name{plot_cv}
\alias{plot_cv}
\title{Plot deconvolution lambda cross-validation curve}
\usage{
plot_cv(fit, ...)
}
\arguments{
\item{fit}{Object of class 'deconv'}

\item{...}{Optional arguments passed to \code{plot()}}
}
\value{
No return value. Plots the lambda CV curve using base graphics.
}
\description{
Plots the cross-validation curve, and upper and lower standard error curves
as a function of the ridge parameter lambda.
}
