% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bulk.R
\name{plot_pred}
\alias{plot_pred}
\title{Scatter plot to compare deconvoluted subclasses}
\usage{
plot_pred(obs, pred, mk = NULL, scheme = NULL, ellipse = NULL)
}
\arguments{
\item{obs}{Observed matrix of cell amounts with subclasses in columns and
samples in rows.}

\item{pred}{Predicted (deconvoluted) matrix of cell amounts with rows and
columns matching \code{obs}.}

\item{mk}{Optional matching cellMarkers object. This is used for its
\code{cell_table} element to try to colour subclasses by group.}

\item{scheme}{Vector of colours, one for each cell subclass.}

\item{ellipse}{Either a single number for the number of ellipses to plot, in
which case ellipses are shown for cell subclasses with the lowest R^2; or a
character vector of cell subclasses to be outlined with ellipses. Requires
the ggforce package to be installed.}
}
\value{
A ggplot2 scatter plot. An overall R^2 (coefficient of
determination) comparing all observed and predicted results is shown.
}
\description{
Produces a single scatter plot using base graphics to compare actual cell
counts against deconvoluted cell counts from bulk (or pseudo-bulk) RNA-Seq
Cell subclasses are shown in different colours. Designed for use if ground
truth is available, e.g. for simulated pseudo-bulk RNA-Seq data.
}
