% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_map.R
\name{quantile_map}
\alias{quantile_map}
\title{Quantile mapping function between two scRNA-Seq datasets}
\usage{
quantile_map(x, y, n = 10000, remove_noncoding = TRUE)
}
\arguments{
\item{x}{scRNA-Seq data whose distribution is to be mapped onto \code{y}: either a
matrix of gene expression on log2+1 scale, or a 'cellMarkers' class object,
in which case the \verb{$genemeans} list element is extracted.}

\item{y}{Reference scRNA-Seq data: either a matrix of gene expression on
log2+1 scale, or a 'cellMarkers' class object, in which case the
\verb{$genemeans} list element is extracted.}

\item{n}{Number of quantiles to split x and y}

\item{remove_noncoding}{Logical, whether to remove noncoding. This is a basic
filter which looks at the gene names (rownames) in both matrices and
removes genes containing "-" which are usually antisense or mitochondrial
genes, or "." which are either pseudogenes or ribosomal genes.}
}
\value{
A function of form \code{FUN(x)} where \code{x} can be supplied as a numeric
vector or matrix and the same type is returned.
}
\description{
Quantile mapping to combine two scRNA-Seq datasets based on mapping either
the distribution of mean log2+1 gene expression in cell clusters to the
distribution of the 2nd dataset, or mapping the quantiles of one matrix of
gene expression (with genes in rows) to another.
}
