% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudent.R
\name{residuals.deconv}
\alias{residuals.deconv}
\title{Extract Deconvolution Residuals}
\usage{
\method{residuals}{deconv}(object, ..., type = c("raw", "weight"), test = NULL)
}
\arguments{
\item{object}{a 'deconv' class object}

\item{...}{retained for class compatibility}

\item{type}{Specifies type of residuals, either raw or weighted. Ignored if
\code{test} is specified.}

\item{test}{bulk gene expression matrix assumed to be in raw counts}
}
\value{
Matrix of residuals.
}
\description{
Extracts residuals from a deconvolution model. As the model uses a reduced
signature gene set for deconvolution, in order to extract residuals for all
genes, these need to recalculated by supplying the bulk count matrix \code{test}.
}
