% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bulk.R
\name{Rsq_set}
\alias{Rsq_set}
\title{Calculate R squared on deconvoluted cell subclasses}
\usage{
Rsq_set(obs, pred, force_intercept = FALSE)
}
\arguments{
\item{obs}{Observed matrix of cell amounts with subclasses in columns and
samples in rows.}

\item{pred}{Predicted (deconvoluted) matrix of cell amounts with rows and
columns matching \code{obs}.}

\item{force_intercept}{Logical whether to force intercept through 0.}
}
\value{
Vector of R squared values calculated using \code{lm()}.
}
\description{
Calculates R squared comparing subclasses in each column of \code{obs} with
matching columns in deconvoluted \code{pred}. Samples are in rows. For use if
ground truth is available, e.g. simulated pseudo-bulk RNA-Seq data.
}
