% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudent.R
\name{rstudent.deconv}
\alias{rstudent.deconv}
\alias{rstandard.deconv}
\alias{cooks.distance.deconv}
\title{Regression Deletion Diagnostics}
\usage{
\method{rstudent}{deconv}(model, ...)

\method{rstandard}{deconv}(model, ...)

\method{cooks.distance}{deconv}(model, ...)
}
\arguments{
\item{model}{'deconv' class object}

\item{...}{retained for class compatibility}
}
\value{
Matrix of adjusted residuals or Cook's distance.
}
\description{
Functions for computing regression diagnostics.
}
\details{
\code{rstandard} and \code{rstudent} give standardized and Studentized residuals
respectively.
}
\seealso{
\code{\link[stats:influence.measures]{stats::influence.measures()}}
}
