% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudent.R
\name{se}
\alias{se}
\title{Standard errors of deconvoluted cell counts}
\usage{
se(model, type = c("var.e", "OLS", "OLS2", "HC0", "HC2", "HC3"))
}
\arguments{
\item{model}{a 'deconv' class object}

\item{type}{specifies standard error method. Default is "var.e", which we
recommend as being the most accurate. The alternative options include OLS
(ordinary least squares) and several heteroscedasticity-consistent methods,
HC0, HC2 and HC3.}
}
\value{
with \code{type = "var.e"}, a vector of standard errors of cell counts
for each cell subclass. For all other options, a matrix of SE for every
bulk sample and every cell subclass.
}
\description{
Extracts standard errors of deconvoluted cell counts based on the row
variance of weighted residuals.
}
\seealso{
\code{\link[=deconvolute]{deconvolute()}}
}
