% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_heatmap.R
\name{signature_heatmap}
\alias{signature_heatmap}
\title{Gene signature heatmap}
\usage{
signature_heatmap(x, ...)
}
\arguments{
\item{x}{Object of S3 class 'cellMarkers' generated by \code{\link[=cellMarkers]{cellMarkers()}}.}

\item{...}{Optional arguments passed to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.}
}
\description{
Produces a heatmap of genes signatures for each cell subclass using
ComplexHeatmap.
}
