% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bulk.R
\name{simulate_bulk}
\alias{simulate_bulk}
\title{Simulate pseudo-bulk RNA-Seq}
\usage{
simulate_bulk(object, sim_counts)
}
\arguments{
\item{object}{A 'cellMarkers' class object}

\item{sim_counts}{An integer matrix with samples in rows and columns for each
cell subclass in \code{object}. This can be generated using \code{\link[=generate_cells]{generate_cells()}}.}
}
\value{
An integer count matrix with genes in rows and cell subclasses in
columns. This can be used as \code{test} with the \code{\link[=deconvolute]{deconvolute()}} function.
}
\description{
Simulates pseudo-bulk RNA-Seq dataset based on a 'cellMarkers' class object
and a matrix of counts for the numbers of cells of each cell subclass.
}
\seealso{
\code{\link[=generate_cells]{generate_cells()}} \code{\link[=deconvolute]{deconvolute()}}
}
