% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scapply.R
\name{slapply}
\alias{slapply}
\title{Apply a function to a big matrix by slicing}
\usage{
slapply(
  x,
  FUN,
  combine = "c",
  big = NULL,
  verbose = TRUE,
  sliceSize = 1000L,
  cores = 1L,
  ...
)
}
\arguments{
\item{x}{matrix or sparse matrix of raw counts with genes in rows and cells
in columns.}

\item{FUN}{Function to be applied to each subblock of the matrix.}

\item{combine}{A function or a name of a function to combine results after
slicing, i.e. only invoked if \code{big} is \code{TRUE}. As the function is usually
applied to blocks of 1000 genes or so, the result is usually a vector wih
an element per gene. Hence 'c' is the default function for combining
vectors into a single longer vector. However if each gene row returns a
number of results (e.g. a vector or dataframe), then \code{combine} could be set
to 'rbind'.}

\item{big}{Logical, whether to invoke slicing of \code{x} into rows. This is
invoked automatically if \code{x} is a large matrix with >2^31 elements.}

\item{verbose}{Logical, whether to show progress.}

\item{sliceSize}{Integer, number of rows of \code{x} to use in each slice if
\code{big = TRUE}.}

\item{cores}{Integer, number of cores to use for parallelisation using
\code{mclapply()}. Parallelisation is not available on windows. Warning:
parallelisation has increased memory requirements.}

\item{...}{Optional arguments passed to \code{FUN}.}
}
\value{
The returned data type will depend on the functions specified by
\code{FUN} and \code{combine}.
}
\description{
Workhorse function ('slice apply') designed to handle large scRNA-Seq gene
expression matrices such as embedded Seurat matrices, and apply a function to
the whole matrix. Very large matrices are handled by slicing rows into blocks
to avoid excess memory requirements.
}
\details{
The limit on \code{sliceSize} is that the number of elements manipulated in each
block (i.e. \code{sliceSize} x number of cells in a given subclass/group) must be
kept below the long vector limit of 2^31 (around 2e9). Increasing \code{cores}
and/or \code{sliceSize} requires substantial amounts of spare RAM. \code{combine} works
in a similar way to \code{.combine} in \code{foreach()} across slices of genes; it
is only invoked if \code{big} is \code{TRUE}.
}
\seealso{
\code{\link[=scapply]{scapply()}}
}
\author{
Myles Lewis
}
