% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_plot.R
\name{specificity_plot}
\alias{specificity_plot}
\title{Specificity plot}
\usage{
specificity_plot(
  mk,
  subclass,
  use_filter = FALSE,
  nrank = 8,
  label_pos = "right",
  nsubclass = 5,
  axis_extend = 0.4,
  scheme = NULL
)
}
\arguments{
\item{mk}{a 'cellMarkers' class object.}

\item{subclass}{character value specifying the subclass to be plotted}

\item{use_filter}{logical, whether to use gene mean expressions to which
noise reduction filtering has been applied.}

\item{nrank}{number of ranks of subclasses to display.}

\item{label_pos}{character value, either "left" or "right" specifying which
side to add labels.}

\item{nsubclass}{number of top markers to label}

\item{axis_extend}{numeric value, specifying how far to extend the x axis to
the left as a proportion. Only invoked when \code{label_pos = "left"}.}

\item{scheme}{Vector of colours for points.}
}
\value{
a ggplot2 scatter plot object.
}
\description{
Scatter plot showing specificity of genes as markers for a particular cell
subclass.
}
