% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_plot.R
\name{stack_plot}
\alias{stack_plot}
\title{Stacked bar plot}
\usage{
stack_plot(
  x,
  percent = FALSE,
  order_col = 1,
  scheme = NULL,
  cex.names = 0.7,
  ...
)
}
\arguments{
\item{x}{matrix of deconvolution results with samples in rows and cell
subclasses or groups in columns.}

\item{percent}{Logical whether to scale the matrix rows as percentage.}

\item{order_col}{Numeric value for which column to sort. If a vector of
column indices is supplied, these columns are averaged first using
\code{rowMeans()}.}

\item{scheme}{Vector of colours. If not supplied, the default scheme uses
\code{scales::hue_pal()}.}

\item{cex.names}{Character expansion controlling bar names font size.}

\item{...}{Optional arguments passed to \code{\link[graphics:barplot]{graphics::barplot()}}.}
}
\value{
No return value. Plots a stacked barchart using base graphics.
}
\description{
Produces stacked bar plots using base graphics.
}
