% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_deconv.R
\name{summary.tune_deconv}
\alias{summary.tune_deconv}
\title{Summarising deconvolution tuning}
\usage{
\method{summary}{tune_deconv}(object, ...)
}
\arguments{
\item{object}{dataframe of class \code{'tune_deconv'}.}

\item{...}{further arguments passed to other methods.}
}
\value{
Prints the row representing the best tuning of parameters (maximum
mean R squared, averaged across subclasses). Invisibly returns a dataframe
of mean R squared values averaged over subclasses.
}
\description{
\code{summary} method for class \code{'tune_deconv'}.
}
