% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMarkers.R
\name{updateMarkers}
\alias{updateMarkers}
\title{Update cellMarkers object}
\usage{
updateMarkers(
  object = NULL,
  genemeans = NULL,
  groupmeans = NULL,
  add_genes = NULL,
  add_groupgenes = NULL,
  nsubclass = 5,
  ngroup = 5,
  expfilter = 1,
  noisefilter = 1.5,
  noisefraction = 0.25,
  verbose = TRUE
)
}
\arguments{
\item{object}{A cellMarkers class object. Either \code{object} or \code{genemeans} must
be specified.}

\item{genemeans}{A matrix of mean gene expression with genes in rows and cell
subclasses in columns.}

\item{groupmeans}{Optional matrix of mean gene expression for overarching
main cell groups (genes in rows, cell groups in columns).}

\item{add_genes}{Character vector of gene markers to add manually to the cell
subclass gene signature.}

\item{add_groupgenes}{Character vector of gene markers to add manually to the
cell group gene signature.}

\item{nsubclass}{Number of genes to select for each single cell subclass.}

\item{ngroup}{Number of genes to select for each cell group.}

\item{expfilter}{Genes whose maximum mean expression on log2 scale per cell
type are below this value are removed.}

\item{noisefilter}{Numeric value below which mean log2 gene expression is
reduced to 0.}

\item{noisefraction}{Numeric value. Maximum mean log2 gene expression across
cell types is calculated and values in celltypes below this fraction are
set to 0.}

\item{verbose}{Logical whether to show messages.}
}
\value{
Returns list object containing \code{best_angle}, a list of genes ranked
by lowest angle and highest maximum expression in a cell type; \code{genemeans},
matrix of mean log2+1 gene expression with genes in rows and cell types in
columns; \code{genemeans_filters}, matrix of gene expression following noise
reduction.
}
\description{
Updates a cellMarkers gene signature object without having to rerun
calculation of gene means which can be slow.
}
