% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_plot.R
\name{violin_plot}
\alias{violin_plot}
\title{Cell subclass violin plot}
\usage{
violin_plot(x, percent = FALSE, order_cols = c("none", "increase", "decrease"))
}
\arguments{
\item{x}{matrix of deconvolution results with samples in rows and cell
subclasses or groups in columns. If a 'deconv' class object is supplied the
deconvolution values for the cell subclasses are extracted and plotted.}

\item{percent}{Logical whether to scale the matrix rows as percentage.}

\item{order_cols}{Character value specifying with cell types are ordered by
mean abundance.}
}
\value{
A ggplot2 plotting object.
}
\description{
Produces violin plots using ggplot2 showing amounts of cell subclasses in
deconvoluted bulk samples.
}
