% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_bulk.R
\name{add_noise}
\alias{add_noise}
\title{Add noise to count data}
\usage{
add_noise(counts, sd = 100, sd2 = 0.05, type = 1)
}
\arguments{
\item{counts}{An integer count matrix with genes in rows and cell
subclasses typically generated by \code{\link[=simulate_bulk]{simulate_bulk()}}.}

\item{sd}{Standard deviation of noise added to counts.}

\item{sd2}{Standard deviation of noise added to log2(counts+1). Only applied
if \code{type = 1}.}

\item{type}{Either \code{1} for Gaussian noise on counts and/or log2 counts. Or
\code{2} in which case Gaussian noise is applied to square root transformed
counts.}
}
\value{
An integer count matrix with genes in rows and cell subclasses in
columns.
}
\description{
Gaussian noise can be added to the simulated count matrix in 2 ways which can
be combined. \code{sd} controls addition of simple Gaussian noise to counts using
\code{rnorm} with sd specified by \code{sd}. If \code{sd2} is >0, counts are converted
using log2+1 and Gaussian noise added, followed by conversion back to count
scale. Negative values are converted to 0. \code{sd} affects low expressed genes
and hardly affects high expressed genes. \code{sd2} affects all genes irrespective
of expression level.
An alternative system of adding noise is invoked if \code{type = 2}. Here counts
are square root transformed before Gaussian noise is added, and then
transformed back. This still has a stronger effect on low expressed genes,
but the effect is more graduated with a more gradual fall off in effect on
genes with increasing expression. \code{sd} is scaled to a mean library size
(total counts in each sample/column) of 1e9 counts to adjust for the effect
of sampling more cells in \code{\link[=simulate_bulk]{simulate_bulk()}} by increasing \code{times}.
}
