% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudent.R
\name{confint.deconv}
\alias{confint.deconv}
\title{Confidence Intervals for Deconvolution Models}
\usage{
\method{confint}{deconv}(object, parm, level = 0.95, ..., type = "var.e")
}
\arguments{
\item{object}{a fitted 'deconv' class model object.}

\item{parm}{for compatibility with S3 method. Not used.}

\item{level}{the confidence level required.}

\item{...}{additional arguments for S3 compatibility.}

\item{type}{specifies standard error method, see \code{\link[=se]{se()}}.}
}
\value{
List containing 2 matrices with lower and upper confidence
intervals.
}
\description{
Computes confidence intervals for fitted deconvolution models.
}
\details{
The default "var.e" method for calculating standard error (SE) is
recommended. Other SE methods are most reliable with bulk data with low
noise, and when fitted compensation values are close to 1 and ridge parameter
lambda is close to 0. As compensation values deviate away from 1 or as lambda
increases, CI calculated based on OLS/HC0-3 SE inflate and are progressively
less reliable.
}
\seealso{
\code{\link[=se]{se()}}
}
