% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute}
\alias{deconvolute}
\title{Deconvolute bulk RNA-Seq using single-cell RNA-Seq signature}
\usage{
deconvolute(
  mk,
  test,
  comp_amount = 0,
  group_comp_amount = 0,
  adjust_comp = TRUE,
  use_filter = TRUE
)
}
\arguments{
\item{mk}{object of class 'cellMarkers'}

\item{test}{matrix of bulk RNA-Seq counts to be deconvoluted}

\item{comp_amount}{either a single value from 0-1 for the amount of
compensation or a numeric vector with the same length as the number of cell
subclasses to deconvolute.}

\item{group_comp_amount}{either a single value from 0-1 for the amount of
compensation for cell group analysis or a numeric vector with the same
length as the number of cell groups to deconvolute.}

\item{adjust_comp}{logical, whether to optimise \code{comp_amount} to prevent
negative cell proportion projections.}

\item{use_filter}{logical, whether to use denoised signature matrix}
}
\value{
a list object containing the cell proportions of each cell subclass
in each sample in element \code{subclass}, and the cell proportions of each cell
group in element \code{group}.
}
\description{
Deconvolution of bulk RNA-Seq using vector projection method with optional
compensation for spillover.
}
