% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute}
\alias{deconvolute}
\title{Deconvolute bulk RNA-Seq using single-cell RNA-Seq signature}
\usage{
deconvolute(
  mk,
  test,
  log = TRUE,
  exp_signature = FALSE,
  comp_amount = 1,
  group_comp_amount = 0,
  equal_weight = FALSE,
  adjust_comp = TRUE,
  use_filter = TRUE,
  arith_mean = FALSE,
  convert_bulk = "ref",
  plot_comp = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mk}{object of class 'cellMarkers'. See \code{\link[=cellMarkers]{cellMarkers()}}.}

\item{test}{matrix of bulk RNA-Seq to be deconvoluted. We recommend raw
counts as input, but normalised data can be provided, in which case set
\code{log = FALSE}.}

\item{log}{Logical, whether to apply log2 +1 to count data in \code{test}. Set to
\code{FALSE} if prenormalised bulk RNA-Seq data is provided.}

\item{exp_signature}{Logical, whether signature and test revert to count
scale by 2^ exponentiation.}

\item{comp_amount}{either a single value from 0-1 for the amount of
compensation or a numeric vector with the same length as the number of cell
subclasses to deconvolute.}

\item{group_comp_amount}{either a single value from 0-1 for the amount of
compensation for cell group analysis or a numeric vector with the same
length as the number of cell groups to deconvolute.}

\item{equal_weight}{logical, whether the gene signature matrix is scaled to
equalise the effects of each gene.}

\item{adjust_comp}{logical, whether to optimise \code{comp_amount} to prevent
negative cell proportion projections.}

\item{use_filter}{logical, whether to use denoised signature matrix.}

\item{arith_mean}{logical, whether to use arithmetic means (if available) for
signature matrix. Mainly useful with pseudo-bulk simulation.}

\item{convert_bulk}{either "ref" to convert bulk RNA-Seq to scRNA-Seq scaling
using reference data or "qqmap" using quantile mapping of the bulk to
scRNA-Seq datasets, or "none" for no conversion.}

\item{plot_comp}{logical, whether to analyse compensation values across
subclasses.}

\item{verbose}{logical, whether to show additional information.}
}
\value{
A list object of S3 class 'deconv' containing:
\item{call}{the matched call}
\item{mk}{the original 'cellMarkers' class object}
\item{subclass}{list object containing:
\itemize{
\item \code{output}, the amount of each subclass based purely on project
gene expression
\item \code{percent}, the proportion of each subclass scaled as a percentage
so that the total amount across all subclasses adds to 100\%
\item \code{spillover}, the spillover matrix
\item \code{compensation}, the mixed final compensation matrix which
incorporates \code{comp_amount}
\item \code{rawcomp}, the original unadjusted compensation matrix
\item \code{comp_amount}, the final values for the amount of compensation
across each cell subclass after adjustment to prevent negative values
}}
\item{group}{similar list object to \code{subclass}, but with results for the
cell group analysis.}
\item{nest_output}{alternative matrix of cell output results for each
subclass adjusted so that the cell outputs across subclasses are nested
as a proportion of cell group outputs.}
\item{nest_percent}{alternative matrix of cell proportion results for each
subclass adjusted so that the percentages across subclasses are nested
within cell group percentages. The total percentage still adds to 100\%.}
\item{comp_amount}{original argument \code{comp_amount}}
\item{comp_check}{optional list element returned when \code{plot_comp = TRUE}}
}
\description{
Deconvolution of bulk RNA-Seq using vector projection method with adjustable
compensation for spillover.
}
\seealso{
\code{\link[=cellMarkers]{cellMarkers()}} \code{\link[=updateMarkers]{updateMarkers()}}
}
\author{
Myles Lewis
}
