% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp.R
\name{plot_comp}
\alias{plot_comp}
\alias{plot_path}
\title{Plot compensation analysis}
\usage{
plot_comp(x, overlay = TRUE, mfrow = NULL, ...)

plot_path(x, subclass = 1L, sample = NULL, add_points = FALSE, ...)
}
\arguments{
\item{x}{An object of class 'deconv' generated by \code{\link[=deconvolute]{deconvolute()}}.}

\item{overlay}{Logical whether to overlay compensation curves onto a single
plot.}

\item{mfrow}{Optional vector of length 2 for organising plot layout. See
\code{par()}. Only used when \code{overlay = FALSE}.}

\item{...}{Optional graphical arguments passed to \code{\link[=plot]{plot()}}.}

\item{subclass}{Either an index number or character value specifying which cell
subclass to plot.}

\item{sample}{Either an index number or character value specifying which
sample to plot. Both \code{subclass} and \code{sample} cannot be specified together.}

\item{add_points}{Logical whether to add points showing the final cell count
values.}
}
\value{
No return value. \code{plot_comp()} plots the effect of varying
compensation on the minimum subclass output for every cell subclass.
\code{plot_path()} plots the coefficient paths for either each bulk sample for the
subclass specified by \code{subclass}, or all subclasses for a single sample
specified by \code{sample}.
}
\description{
Plots the effect of varying compensation from 0 to 1 for each cell subclass,
examining the minimum subclass output result following a call to
\code{\link[=deconvolute]{deconvolute()}}. For this function to work, the argument \code{plot_comp} must be
set to \code{TRUE} during the call to \code{\link[=deconvolute]{deconvolute()}}.
}
