% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_deconv.R
\name{plot_tune}
\alias{plot_tune}
\title{Plot tuning curves}
\usage{
plot_tune(res, group = "subclass", xvar = colnames(res)[1], title = "")
}
\arguments{
\item{res}{Dataframe of tuning results generated by \code{\link[=tune_deconv]{tune_deconv()}}.}

\item{group}{Character value specifying column in \code{res} to be grouped by
colour.}

\item{xvar}{Character value specifying column in \code{res} to vary along the x
axis.}

\item{title}{Character value for the plot title.}
}
\value{
ggplot2 scatter plot.
}
\description{
Produces a ggplot2 plot of R-squared values generated by \code{\link[=tune_deconv]{tune_deconv()}}.
}
