% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_deconv.R
\name{plot_tune}
\alias{plot_tune}
\title{Plot tuning curves}
\usage{
plot_tune(result, group = "subclass", xvar = colnames(result)[1], title = "")
}
\arguments{
\item{result}{Dataframe of tuning results generated by \code{\link[=tune_deconv]{tune_deconv()}}.}

\item{group}{Character value specifying column in \code{result} to be grouped by
colour.}

\item{xvar}{Character value specifying column in \code{result} to vary along the x
axis.}

\item{title}{Character value for the plot title.}
}
\value{
ggplot2 scatter plot.
}
\description{
Produces a ggplot2 plot of R-squared values generated by \code{\link[=tune_deconv]{tune_deconv()}}.
}
\details{
If \code{group} is set to \code{"subclass"}, then the tuning parameter specified by
\code{xvar} is varied on the x axis. Any other tuning parameters (i.e. if 2 or
more have been tuned) are fixed to their best tuned values.

If \code{group} is set to a different column than \code{"subclass"}, then the mean
R-squared values in \code{result} are averaged over subclasses. This makes it
easier to compare the overall effect (mean R-squared) of 2 tuned parameters
which are specified by \code{xvar} and \code{group}. Any remaining parameters not shown
are fixed to their best tuned values.
}
