% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudent.R
\name{residuals.deconv}
\alias{residuals.deconv}
\title{Extract Deconvolution Residuals}
\usage{
\method{residuals}{deconv}(object, test = NULL, arith_mean = FALSE, use_filter = FALSE, ...)
}
\arguments{
\item{object}{a 'deconv' class object}

\item{test}{bulk gene expression matrix assumed to be in raw counts}

\item{arith_mean}{logical, whether to use arithmetic mean as gene signature}

\item{use_filter}{logical, whether to use denoised signature matrix}

\item{...}{retained for class compatibility}
}
\value{
Matrix of residuals.
}
\description{
Extracts residuals from a deconvolution model. As the model uses a reduced
signature gene set for deconvolution, in order to extract residuals for all
genes, these need to recalculated by supplying the bulk count matrix \code{test}.
}
