% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmean.R
\name{scmean}
\alias{scmean}
\title{Single-cell mean log gene expression across cell types}
\usage{
scmean(x, celltype, big = NULL, verbose = TRUE)
}
\arguments{
\item{x}{matrix or sparse matrix of raw counts with genes in rows and cells
in columns.}

\item{celltype}{a vector of cell subclasses or types whose length matches the
number of columns in \code{x}. It is coerced to a factor. \code{NA} are tolerated and
the matching columns in \code{x} are skipped.}

\item{big}{Logical, whether to invoke slicing of \code{x} into rows. This is
invoked automatically if \code{x} is a large matrix with >2^31 elements.}

\item{verbose}{Logical, whether to print messages.}
}
\value{
a matrix of mean log2 gene expression across cell types with genes
in rows and cell types in columns.
}
\description{
Workhorse function which takes as input a scRNA-Seq gene expression matrix
such as embedded in a Seurat object, calculates log2(counts +1) and averages
gene expression over a vector specifying cell subclasses or cell types. Very
large matrices are handled by slicing rows into blocks to avoid excess memory
requirements.
}
