% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudent.R
\name{se}
\alias{se}
\title{Standard errors of deconvoluted cell counts}
\usage{
se(model)
}
\arguments{
\item{model}{a 'deconv' class object}
}
\value{
a vector of the standard errors of cell counts for each cell
subclass
}
\description{
Extracts standard errors of deconvoluted cell counts based on the row
variance of weighted residuals.
}
\seealso{
\code{\link[=deconvolute]{deconvolute()}}
}
