% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_plot.R
\name{stack_plot}
\alias{stack_plot}
\alias{stack_ggplot}
\title{Stacked bar plot}
\usage{
stack_plot(
  x,
  percent = FALSE,
  order_col = 1,
  scheme = NULL,
  order_cells = c("none", "increase", "decrease"),
  cex.names = 0.7,
  ...
)

stack_ggplot(
  x,
  percent = FALSE,
  order_col = 1,
  scheme = NULL,
  order_cells = c("none", "increase", "decrease"),
  legend_ncol = NULL,
  legend_position = "bottom"
)
}
\arguments{
\item{x}{matrix of deconvolution results with samples in rows and cell
subclasses or groups in columns. If a 'deconv' class object is supplied the
deconvolution values for the cell subclasses are extracted and plotted.}

\item{percent}{Logical whether to scale the matrix rows as percentage.}

\item{order_col}{Numeric value for which column to sort. If a vector of
column indices is supplied, these columns are averaged first using
\code{rowMeans()}.}

\item{scheme}{Vector of colours. If not supplied, the default scheme uses
\code{scales::hue_pal()}.}

\item{order_cells}{Character value specifying with cell types are ordered by
abundance.}

\item{cex.names}{Character expansion controlling bar names font size.}

\item{...}{Optional arguments passed to \code{\link[graphics:barplot]{graphics::barplot()}}.}

\item{legend_ncol}{Number of columns for ggplot2 legend. If set to \code{NULL}
ggplot2 sets the column number automatically.}

\item{legend_position}{Position of ggplot2 legend}
}
\value{
No return value. Plots a stacked barchart using base graphics.
}
\description{
Produces stacked bar plots using base graphics or ggplot2.
}
