% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMarkers.R
\name{updateMarkers}
\alias{updateMarkers}
\title{Update cellMarkers object}
\usage{
updateMarkers(
  object = NULL,
  genemeans = NULL,
  groupmeans = NULL,
  add_genes = NULL,
  add_groupgenes = NULL,
  bulkdata = NULL,
  nsubclass = 5,
  ngroup = 5,
  expfilter = 1,
  noisefilter = 2,
  noisefraction = 0.25,
  verbose = TRUE
)
}
\arguments{
\item{object}{A cellMarkers class object. Either \code{object} or \code{genemeans} must
be specified.}

\item{genemeans}{A matrix of mean gene expression with genes in rows and cell
subclasses in columns.}

\item{groupmeans}{Optional matrix of mean gene expression for overarching
main cell groups (genes in rows, cell groups in columns).}

\item{add_genes}{Character vector of gene markers to add manually to the cell
subclass gene signature.}

\item{add_groupgenes}{Character vector of gene markers to add manually to the
cell group gene signature.}

\item{bulkdata}{Optional data matrix containing bulk RNA-Seq data with genes
in rows. This matrix is only used for its rownames, to ensure that cell
markers are selected from genes in the bulk dataset.}

\item{nsubclass}{Number of genes to select for each single cell subclass.}

\item{ngroup}{Number of genes to select for each cell group.}

\item{expfilter}{Genes whose maximum mean expression on log2 scale per cell
type are below this value are removed and not considered for the signature.}

\item{noisefilter}{Sets an upper bound for \code{noisefraction} cut-off below
which gene expression is set to 0. Essentially gene expression above this
level must be retained in the signature. Setting this higher can allow more
suppression via \code{noisefraction} and can favour more highly expressed genes.}

\item{noisefraction}{Numeric value. Maximum mean log2 gene expression across
cell types is calculated and values in celltypes below this fraction are
set to 0. Set in conjunction with \code{noisefilter.} Note: if this is set too
high (too close to 1), it can have a deleterious effect on deconvolution.}

\item{verbose}{Logical whether to show messages.}
}
\value{
Returns list object containing \code{best_angle}, a list of genes ranked
by lowest angle and highest maximum expression in a cell type; \code{genemeans},
matrix of mean log2+1 gene expression with genes in rows and cell types in
columns; \code{genemeans_filters}, matrix of gene expression following noise
reduction.
}
\description{
Updates a cellMarkers gene signature object without having to rerun
calculation of gene means which can be slow.
}
